create price card
