+++ Code/Client/users/src/components/molecules/carRental/priceCard/rentalPriceCard.module.css
@@ -1,0 +1,39 @@
+.contanier{
+    width: 20rem;
+}
+
+.card{
+    overflow: hidden;
+    padding: 1.5rem 1.8rem;
+    background-color: white;
+    box-shadow: 0 2px 8px 2px rgba(178, 178, 178, 0.189);
+  }
+
+.header{
+    border-bottom: 0.1rem solid rgb(182, 179, 179);
+    padding-bottom: 10px;
+}
+
+.body{
+    padding-top: 10px;
+    border-bottom: 0.1rem solid rgb(182, 179, 179);
+    padding-bottom: 10px;
+}
+
+.icon{
+    font-size: 1.4rem;
+}
+
+.price{
+    padding-top: 10px;
+    display: flex;
+    justify-content: space-between;
+    font-size: 1.2rem;
+    font-weight: 700;
+}
+
+.btn{
+    display: flex;
+    justify-content: flex-end;
+    padding: 10px 0px;
+}
