+++ Code/Client/users/src/pages/CarRental/addOnPage/ExtendedDetail.module.css
@@ -1,3 +1,16 @@
-.container{
-    background-color: rgba(235, 235, 240, 1)
-}
+.container {
+    display: flex;
+    flex-direction: row;
+    padding: 60px 40px;
+    background-color: rgb(226, 222, 222);
+    height: 40rem;
+    gap: 40px;
+  }
+  
+  .addOnContainer {
+    width: 80rem;
+    display: flex;
+    flex-wrap: wrap; 
+    gap: 20px;
+    
+  }
