+++ Code/Client/users/src/pages/CarRental/showCarsPage/ViewCars.js
@@ -7,18 +7,18 @@ import style from "./viewCars.module.css"
 export default function ViewCars() {
     return (
         <>
-        <div>
-            <Navbar />
-            <div className={style.container}>
-                <div className={style.filterContainer}>
-                    <FiltersCard />
+            <div>
+                <Navbar />
+                <div className={style.container}>
+                    <div className={style.filterContainer}>
+                        <FiltersCard />
+                    </div>
+                    <div className={style.carContainer}>
+                        <CarCard />
+                        <CarCard />
+                        <CarCard />
+                    </div>
                 </div>
-                <div className={style.carContainer}>
-                    <CarCard />
-                    <CarCard />
-                    <CarCard />
-                </div>
-            </div>
             </div>
         </>
     )
